refactoring, documentation
