+++ config.py
@@ -8,5 +8,12 @@ CONFIG_DIR = os.path.abspath(os.path.join(__file__, os.pardir))
 PATH_DATA_DIR = os.path.join(CONFIG_DIR, 'data')
 os.makedirs(PATH_DATA_DIR, exist_ok=True)
 
+# Key used for the location index
+KEY_LOC = 'loc_id'
+# Key used for the year index
+KEY_YEAR = 'year'
+# Key used for yield targets
+KEY_TARGET = 'yield'
+
