+++ datasets/dataset_torch.py
@@ -1,6 +1,8 @@
 import torch
 import torch.utils.data
 
+from config import KEY_LOC, KEY_YEAR, KEY_TARGET
+
 from datasets.dataset import Dataset
 from util.torch import batch_tensors
 
@@ -21,19 +23,35 @@ class TorchDataset(torch.utils.data.Dataset):
     @classmethod
     def _cast_to_tensor(cls, sample: dict) -> dict:
         return {
-            **{key: sample[key] for key in Dataset.INDEX_KEYS},
-            **{key: torch.tensor(sample[key]) for key in sample.keys() if key not in Dataset.INDEX_KEYS}
+            KEY_LOC: sample[KEY_LOC],
+            KEY_YEAR: sample[KEY_YEAR],
+            KEY_TARGET: torch.tensor(sample[KEY_TARGET]),
+            **{
+                key: torch.tensor(sample[key]) for key in sample.keys() if key not in [KEY_LOC, KEY_YEAR, KEY_TARGET]
+              }  # TODO -- support nonnumeric data?
         }
 
     @classmethod
     def collate_fn(cls, samples: list) -> dict:
-        key_y = Dataset.TARGET_KEY
+        assert len(samples) > 0
+
+        feature_names = set.intersection(*[set(sample.keys()) for sample in samples])
+        feature_names.remove(KEY_TARGET)
+        feature_names.remove(KEY_LOC)
+        feature_names.remove(KEY_YEAR)
 
         batched_samples = {
-            **{key: [sample[key] for sample in samples] for key in Dataset.INDEX_KEYS},
-            **{key: batch_tensors(*[sample[key] for sample in samples]) for key in Dataset.FEATURE_KEYS},
-            key_y: batch_tensors(*[sample[key_y] for sample in samples]),
+            KEY_TARGET: batch_tensors(*[sample[KEY_TARGET] for sample in samples]),
+            KEY_LOC: [sample[KEY_LOC] for sample in samples],
+            KEY_YEAR: [sample[KEY_YEAR] for sample in samples],
+            **{key: batch_tensors(*[sample[key] for sample in samples]) for key in feature_names}
         }
+
+        # batched_samples = {
+        #     **{key: [sample[key] for sample in samples] for key in Dataset.INDEX_KEYS},
+        #     **{key: batch_tensors(*[sample[key] for sample in samples]) for key in Dataset.FEATURE_KEYS},
+        #     key_y: batch_tensors(*[sample[key_y] for sample in samples]),
+        # }
 
         return batched_samples
 
@@ -51,17 +69,12 @@ class TorchDataset(torch.utils.data.Dataset):
 
 if __name__ == '__main__':
 
-    _dataset = Dataset()
-
-    # print(_dataset.years)
-    # print(_dataset.locations)
-    print(_dataset['AL_LAWRENCE', 2000])
-
-    _dataset_train, _dataset_test = Dataset.train_test_datasets()
+    _dataset_train, _dataset_test = Dataset.get_datasets()
 
     _dataset_train_torch = TorchDataset(_dataset_train)
 
-    print(_dataset_train_torch['AL_LAWRENCE', 2000])
+    # print(_dataset_train_torch['AL_LAWRENCE', 2000])
+    print(_dataset_train_torch[1])
 
     import torch.utils.data
     _dataloader = torch.utils.data.DataLoader(
@@ -74,7 +87,7 @@ if __name__ == '__main__':
     for _batch in _dataloader:
 
         print(_batch)
-        print(_batch['YIELD'].shape)
+        print(_batch['yield'].shape)
         print(_batch['TMAX'].shape)
 
         break
